refactor(bench): data-plane split, policy v2, runtimes panel
TUI relocation per operator direction: frontends are forge product, the kit
keeps the data plane. lib/bench/events.py now owns the protocol, demo stream,
ledger adapter and conformance overlay (AST-extracted); viewer/report/
dashboard consume it; tui.py is a pointer stub to forge cli/forge-tui.

Tool policy v2: capabilities are provider registries (browser: harness /
agent-browser / lightpanda / playwright / browserbase; computer use: macos
ladder / computer-use MCP / peekaboo / e2b) with a preferred-provider select,
per-provider allow/ask/deny, add/remove custom rules in the page, and editable
kit.toml keys exporting a per-project .kit.toml override. The hook normalizes
v2 and legacy v1 policies; both paths verified (v2 ask names the preferred
rung, v1 deny still exits 2).

Runtimes panel: detects agent runtimes on the host with real store stats
(6/6 here: claude-code 1763 sessions live-adapted; codex/pi/opencode/gemini/
cursor detect-only, one adapter away, codex first since its rollouts are
JSONL). Answers where other-LLM interactions render: the event protocol is
runtime-neutral; adapters normalize into it.

Proof addendum in docs/verification/observe-agent-surface.md. 5 suites green.
Gate rid policy-v2-runtimes.
